spec-365: split SpecList.tsx components + lift DnD into useSpecBoard [P3]
Behaviour-preserving refactor (audit spec-345 finding sol-6, umbrella
spec-354). SpecList.tsx was a ~985 LOC god-file mixing the KanbanColumn
presentational component, the drag-and-drop state machine, data fetching,
demo reveal, and board actions.

- Extract KanbanColumn (+ AssigneeAvatars, personLabel/initials helpers,
  SpecKanbanStatus/ActiveStatus types) into components/spec-board/.
- Lift the DnD state machine (draggingId/dragOverColumn + start/end/over/
  drop handlers, optimistic updateDocStatus with rollback, board.phase_drag
  telemetry, canWrite guard, Done auto-expand) into a useSpecBoard hook.
- SpecList keeps its public export, data fetch, demo reveal, actions, and
  layout; now 633 LOC. Markup and runtime behaviour are identical.

Verify: tsc -b clean; make test-ui (213 files, 1508 passed); vite build
ok; e2e cold-DB board journeys green (journey-20 kanban phase moves,
journey-29 Done-tab, journey-43 spec.card_opened telemetry — 9 tests).
