+++ packages/ui/src/components/spec-board/types.ts
@@ -1,0 +1,6 @@
+// Per dec-3 / dec-4 of doc-10 the Spec lifecycle is `draft → specify → build →
+// verify → done`. Kanban renders the four active columns; `done` lives in a
+// collapsible rail on the right (dec-5). `approved` is execution-plan-only
+// (t-20 W-B) and never appears on a spec card.
+export type SpecKanbanStatus = 'draft' | 'specify' | 'build' | 'verify' | 'done';
+export type ActiveStatus = Exclude<SpecKanbanStatus, 'done'>;
